requirements.txt
langchain
langchain-openai
langchain-community
librosa
jupyter
openai
pandas
elevenlabs
gradio
python-dotenv
streamlit
pypdf
black
isort